Working combo and equip system!!!!
Attacks now all function as long as they are equipped and the game will successfully find and use the attack. It will also determine if the attack is a finisher or not, and will call nearly anything from the state machine as long as it is equipped and named properly. The developer can also customize the duration, cancel windows, and gravity fields for the animation to play at the specified time.